Split Scheduler Info in presentation and render logic
